Apply review: float64 for the last norm, safer tests, harder CI
All four points from the review, each correct on inspection:

- `_last_update_norm` was the one norm left in float32 when the rest moved
  to float64. In exactly the regime the overflow guard exists for, it would
  report inf while nonfinite was False -- a misleading diagnostic on an
  otherwise healthy clipped run. My oversight.

- The closed-form Conv1D check compared torch.addmm against matmul-then-add
  with torch.equal. Those are different kernels and not guaranteed
  bit-identical across backends or thread counts, which is the same
  nondeterminism class this file documents for C0. Now allclose. The
  toy-vs-HuggingFace comparison stays bit-exact, because both run the same
  addmm on the same shapes and any difference there is a real divergence --
  that is the claim the toy's place in the fast suite rests on.

- Two tests mutating the session-scoped GPT-2 reverted only after their
  assertions, so a failure partway would have handed every later test a
  modified model. Now try/finally, matching TestCeiling.

- CI: persist-credentials: false on both checkouts (neither job pushes, both
  run pip installs), and timeout-minutes on both, so a stalled download
  cannot sit on the 6-hour default.

85 passed fast, 19 passed slow.

+++ plasticity.py
@@ -296,7 +296,7 @@ class OjaPlasticity:
             self.clipped = True
 
         self.delta = new_delta
-        self._last_update_norm = step.norm().item()
+        self._last_update_norm = step.double().norm().item()
         self._write_weight()
         self.n_applied += 1
         return self.report()
